Intermediate state, just updating niceview

+++ app/core/project.py
@@ -1,3 +1,4 @@
+from collections.abc import Iterator
 import datetime
 import os
 from pathlib import Path
@@ -9,6 +10,7 @@ from app.config import app_config
 from app.core.auth import validate_token
 from app.core.models import Project
 from app.util import logger, is_valid_filename
+from niceview.dataadapter import ModelDataAdapter
 
 ###############################################################################
 
@@ -129,6 +131,30 @@ def update_project(project: Project) -> Project:
     return project
 
 
+def rename_project(old_project_name: str, new_project_name: str) -> Project:
+    """
+    Rename a project.
+
+    :param old_project_name: The current name of the project.
+    :param new_project_name: The new name for the project.
+    :return: The updated project object.
+    :raises HTTPException: If the project path is invalid (400 Bad Request).
+    :raises HTTPException: If the new project name already exists (409 Conflict).
+    """
+    old_project_path = get_project_path(old_project_name)
+    new_project_path = get_project_path(new_project_name, check_project_exists=False)
+
+    if new_project_path.exists():
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=f"Project {new_project_name} already exists.")
+
+    try:
+        old_project_path.rename(new_project_path)
+        return get_project(new_project_name)
+    except Exception as e:
+        logger.error(f"Error renaming project {old_project_path} to {new_project_path}: {str(e)}")
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"Error renaming project: {str(e)}")
+
+
 def delete_project(project_name: str) -> None:
     """
     Delete a project.
@@ -138,6 +164,8 @@ def delete_project(project_name: str) -> None:
     """
     project_path = get_project_path(project_name)
     try:
+        if not project_path.exists():
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Project {project_name} does not exist.")
         shutil.rmtree(project_path)
     except Exception as e:
         logger.error(f"Error deleting project {project_path}: {str(e)}")
@@ -190,4 +218,58 @@ def get_auth_project(project_name: str, provisioning_token: str) -> Project:
     project = update_project(project)
 
     return project
+
+###################################################################################
+# Provide operations on the Project model in a nice interface for the UI
+###################################################################################
+
+class ProjectModelAdapter(ModelDataAdapter[Project]):
+    """
+    An adapter for a list of projects to work with the ModelGrid.
+    """
+    def __init__(self) -> None:
+        pass
+
+    def __iter__(self) -> Iterator[Project]:
+        projects = get_projects()
+        return iter(projects)
+
+    def query_all_strs(self) -> Iterator[tuple[str, str]]:
+        projects = get_projects()
+        for index, item in enumerate(projects):
+            yield self.key_from_item(item, index), item.name
+
+    def key_from_item(self, item: Project, index: int = -1) -> str:
+        key = item.name
+        return key
+
+    def key_from_str(self, key: str | int) -> str:
+        return str(key)
+
+    def create(self, item: Project) -> Project:
+        if not isinstance(item, Project):
+            raise TypeError(f"Expected item to be an instance of 'Project', got '{type(item)}'")
+        create_project(item)
+        return item
+
+    def read(self, key: str | int) -> Project:
+        if not isinstance(key, str):
+            raise TypeError(f"Expected key to be a string, got {type(key)}")
+        project = get_project(key)
+        return project
+
+    def update(self, item: Project, key: str) -> Project:
+        if not isinstance(item, Project):
+            raise TypeError(f"Expected item to be an instance of 'Project', got '{type(item)}'")
+        if not isinstance(key, str):
+            raise TypeError(f"Expected key to be a string, got {type(key)}")
+        if key != item.name:
+           rename_project(old_project_name=key, new_project_name=item.name)
+        project = update_project(item)
+        return project
+
+    def delete(self, key: str) -> None:
+        if not isinstance(key, str):
+            raise TypeError(f"Expected key to be a string, got {type(key)}")
+        delete_project(key)
